+++ Gruntfile.js
@@ -6,11 +6,21 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-contrib-clean');
   grunt.loadNpmTasks('grunt-contrib-cssmin');
   grunt.loadNpmTasks('grunt-contrib-watch');
+  grunt.loadNpmTasks('grunt-contrib-copy');
 
+  // The glob does not descend, so src/js/locales/*.js stays out of the main
+  // bundle and is built to dist/locales/ one file at a time instead.
   var jsFiles = [
     'src/js/jquery.grideditor.js',
     'src/js/*.js',
   ];
+
+  var localeFiles = [{
+    expand: true,
+    cwd: 'src/js/locales/',
+    src: ['*.js'],
+    dest: 'dist/locales/',
+  }];
   
   grunt.initConfig({
     pkg: grunt.file.readJSON('package.json'),
@@ -29,7 +39,23 @@ module.exports = function(grunt) {
         },
         src: jsFiles,
         dest: 'dist/jquery.grideditor.min.js',
-      }
+      },
+      locales: {
+        options: {
+          sourceMap: true,
+        },
+        files: localeFiles.map(function(files) {
+          // extDot last, or grideditor.es.js would minify to grideditor.min.js
+          return Object.assign({}, files, { ext: '.min.js', extDot: 'last' });
+        }),
+      },
+    },
+    
+    copy: {
+      // The readable file is what a page loads, next to the minified one
+      locales: {
+        files: localeFiles,
+      },
     },
     
     less: {
@@ -57,7 +83,15 @@ module.exports = function(grunt) {
     watch: {
       stylesheets: {
         files: ['src/**/*', 'example/*'],
-        tasks: ['concat:js', 'uglify', 'less'],
+        tasks: ['concat:js', 'uglify:build', 'less'],
+        options: {
+          spawn: false,
+          livereload: true,
+        },
+      },
+      locales: {
+        files: ['src/js/locales/*.js'],
+        tasks: ['copy:locales', 'uglify:locales'],
         options: {
           spawn: false,
           livereload: true,
@@ -67,6 +101,6 @@ module.exports = function(grunt) {
     
   });
 
-  grunt.registerTask('default', ['concat:js', 'uglify', 'less']);
+  grunt.registerTask('default', ['concat:js', 'uglify', 'less', 'copy:locales']);
 
 };